refactor: merge EmployeeControllers (#22)
